+++ src/client/styles.scss
@@ -1,4 +1,2 @@
-@use "styles/common.scss";
+@use "../common/common.scss";
 @use "styles/main.scss";
-@use "styles/error.scss";
-@use "styles/character.scss";
