+++ src/client/styles/main.scss
@@ -1,4 +1,4 @@
-@use "common.scss";
+@use "../../common/common.scss";
 
 .header {
   background-color: lightgray;
