+++ src/common/components/character-card/character_card.html
@@ -1,6 +1,6 @@
 <div class="character">
   <div class="char-wrapper">
-    <img class="char-portrait" />
+    <img alt="" class="char-portrait" />
     <div class="char-characteristics">
       <div class="char-heading">
         <div class="char-title-wrapper">
