+++ src/common/components/character-card/character_card.scss
@@ -1,4 +1,4 @@
-@use "common.scss";
+@use "../../common.scss";
 
 .character {
   display: flex;
