+++ src/common/components/req-error/req_error.scss
