+++ src/common/html.d.ts
