+++ src/common/models/character.ts
